Added several changes to video downloading functions
Based on flkals' version 3.2 (which uses extractGraphQlMp4Video), I added the ability to pull tweets that are replies. If a tweet cannot be pulled from the home timeline page, the code will now open a new window consisting of the tweet's individual timeline page.

Furthermore, I implemented the ability to pull videos and pictures simultaneously. The code now gathers a videoIndex to locate different videos properly.